# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen produk dan kategori toko",
        "title": "API Meisha Alumunium Kaca",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ari Dev Team",
            "email": "support@meishaalumuniumkaca.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Retrieve all orders with user and product details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order, save user if not exists, and send notification via WebSocket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order and notify via WebSocket",
                "parameters": [
                    {
                        "description": "Order Input",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.OrderInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.OrderInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Mawar No. 123"
                },
                "company_name": {
                    "type": "string",
                    "example": "PT. ABC"
                },
                "details": {
                    "type": "string",
                    "example": "Pesanan baru dari Budi"
                },
                "priority": {
                    "type": "string",
                    "example": "normal"
                },
                "product_id": {
                    "type": "string",
                    "example": "4cd64505-f56f-421b-9243-e6bc9cbbaa7b"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string",
                            "example": "Jl. Mawar No. 123"
                        },
                        "email": {
                            "type": "string",
                            "example": "budi@email.com"
                        },
                        "id": {
                            "type": "string",
                            "example": "4cd64505-f56f-421b-9243-e6bc9cbbaa7b"
                        },
                        "name": {
                            "type": "string",
                            "example": "Budi"
                        },
                        "phone": {
                            "type": "string",
                            "example": "08123456789"
                        }
                    }
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "delivery address",
                    "type": "string"
                },
                "company_name": {
                    "description": "name of the product",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "description": "additional details about the order",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "priority": {
                    "description": "e.g., \"low\", \"normal\", \"high\"",
                    "type": "string"
                },
                "product": {
                    "description": "optional preload",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Product"
                        }
                    ]
                },
                "product_id": {
                    "description": "FK to Product",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "description": "e.g., \"pending\", \"completed\",",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "optional preload",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "UserID and ProductID are foreign keys to User and Product models",
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "addreess": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "langitude": {
                    "description": "bisa null",
                    "type": "string"
                },
                "latitude": {
                    "description": "bisa null",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "optional data"
                },
                "endpoint": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "endpoint": {
                    "type": "string"
                },
                "message": {},
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}